move src/read_until.rs -> src/io/read_until.rs
